aaaaaaaaaa

+++ jpa/src/main/java/br/edu/ufersa/CatCaffe/models/dtos/ProdutoRecordDto.java
@@ -1,8 +1,10 @@
 package br.edu.ufersa.CatCaffe.models.dtos;
 
+import java.math.BigDecimal;
+
 public record ProdutoRecordDto(Long id_produto,
                                String nome_produto,
-                               long preco,
+                               BigDecimal preco,
                                String item,
                                String adicional,
                                Long id_compra) {
